Move site/stages/constants to parent directory
